# docs/swagger.yaml
openapi: "3.0.3"
info:
  title: ISP Billing Platform API
  description: |
    Multi-tenant SaaS backend for ISP billing, FTTH network management, and TR-069 device provisioning.
    All protected endpoints require a valid JWT Bearer token in the Authorization header.
    Tenant isolation is enforced automatically via JWT claims.
  version: "1.0.0"
  contact:
    name: API Support
servers:
  - url: /api/v1
    description: API v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Message:
      type: object
      properties:
        message:
          type: string
...
    Customer:
      type: object
      properties:
        id: { type: string }
        tenant_id: { type: string }
        customer_code: { type: string }
        name: { type: string }
        phone: { type: string }
        email: { type: string }
        address: { type: string }
        latitude: { type: number, nullable: true }
        longitude: { type: number, nullable: true }
        connection_type: { type: string }
        pppoe_username: { type: string }
        pppoe_password: { type: string }
        ip_address: { type: string }
        package_id: { type: string, nullable: true }
        router_id: { type: string, nullable: true }
        odp_port_id: { type: string, nullable: true }
        join_date: { type: string, format: date-time }
        billing_date: { type: integer }
        custom_price: { type: integer, nullable: true }
        discount: { type: integer }
        additional_fee: { type: integer }
        fee_description: { type: string }
        status: { type: string }
        isolated_at: { type: string, format: date-time, nullable: true }
        notes: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        package: { $ref: "#/components/schemas/Package" }
        router: { $ref: "#/components/schemas/Router" }
...
    Package:
      type: object
      properties:
        id: { type: string }
        tenant_id: { type: string }
        name: { type: string }
        description: { type: string }
        bandwidth_up: { type: integer }
        bandwidth_down: { type: integer }
        price: { type: integer }
        burst_limit: { type: string }
        address_list: { type: string }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Router:
      type: object
      properties:
        id: { type: string }
        tenant_id: { type: string }
        name: { type: string }
        identity: { type: string }
        vpn_ip: { type: string }
        vpn_public_key: { type: string }
        radius_secret: { type: string }
        coa_port: { type: integer }
        heartbeat_token: { type: string }
        is_online: { type: boolean }
        last_seen_at: { type: string, format: date-time, nullable: true }
        router_os_ver: { type: string, nullable: true }
        board_name: { type: string, nullable: true }
        uptime: { type: string, nullable: true }
        cpu_load: { type: integer, nullable: true }
        free_memory: { type: integer, nullable: true }
        total_memory: { type: integer, nullable: true }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    ODPPort:
      type: object
      properties:
        id: { type: string }
        odp_id: { type: string }
        port_number: { type: integer }
        customer_id: { type: string, nullable: true }
        status: { type: string }
        notes: { type: string, nullable: true }
        customer: { $ref: "#/components/schemas/Customer" }
...
  parameters:
...
    IDPath:
      name: id
      in: path
      required: true
      schema: { type: string }
...
security:
  - BearerAuth: []
paths:
...
  /odps/{id}/ports/{portId}:
    put:
      tags: [ODPs]
      summary: Update ODP port
      parameters:
        - { $ref: "#/components/parameters/IDPath" }
        - { name: portId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customer_id: { type: string }
                status: { type: string }
                notes: { type: string }
      responses:
        "200":
          description: ODP port updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ODPPort" }
    delete:
      tags: [ODPs]
      summary: Delete ODP port
      parameters:
        - { $ref: "#/components/parameters/IDPath" }
        - { name: portId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }

  # ── Splitters ──
...